ExitCovariant, ExitFailureCovariant, FiberCovariant
